Refactored the scripts into a python module
