+++ skema/text_reading/mention_linking/data/gromet/bucky_simplified_v1--Gromet-FN-auto.json
